BUG: py3: Fix iteration error in GLRLM
In python 3.5 GLRLM results differed from baseline. This was due to
the different implementation of the `filter()` function, which returns
an iterator instead of a list. In GLRLM calculation, there where 2
iterations over this iter-object (one for flat-angle check, one for
matrix calculation), causing it to break in python 3.5.

Fix by merging the 2 iterations into one loop.

Additionally, simplify the filter function.