STYL: Parameterize enabling of C extensions
Add a new parameter ('enableCExtensions') to the settings that can be specified in the parameter file or via `kwargs` at initialization of `featureextractor`. Add this parameter also to the parameter file schema.
Apply this upon execute, as nothing happens if setting is not changed, but is checked regularly.

Finally, fix a small bug, where a incorrect warning is logged when `enableCExtensions` is called when extensions are already enabled. In this case, nothing should happen.