ENH: If enabled, remove outliers after resampling
If `removeOutliers` is set to a float value > 0, outliers (values differing more than n*(standard deviation) from the mean) are set to the outlier value, and n is equal to the value of this parameter. If set to None, no outliers are removed. This is done after image is normalized (i.e. the outliers contribute to the calculation of the mean and standard deviation), but before the scale is applied (i.e. the range of the entire image is between -n*scale and n*scale).