MATH: Revise Compactness 1
Flip exponent (2/3 -> 3/2) of Compactness 1 feature formula. The previous exponent was most likely an error, as this resulted in a formula that was not dimensionless, as the description indicates. The formula with the new exponent is dimensionless.

Compactness1, Compactness2, Sphericity and Spherical disproportion are mathematically correlated. Update the documentation to reflect this (provide equations specifying the correlation between the different features). Also disable Compactness 1 and Compactness 2 in the default/example parameter file in the bin folder to emphasize the redundancy of these features.

Updated the formulas and documentation thereof to have simplified formulas where possible.