DEL: Remove Roundness Feature
Remove the Roundness Feature, as this expresses the same formula as Sphericity, but is calculated by Simple ITK's LabelShapeStatisticsImageFilter (LSSIF). It is documented here: http://www.insight-journal.org/browse/publication/176.
The calculated values differ however, but this is due to the fact that SimpleITK calculates the surface area differently. The surface area calculated by SimpleITK can be found by calling LSSIF.GetPerimeter().

Calculation of Roundness and Surface Area is not done using SimpleITK, as the calculated surface area can be smaller than the surface area of a sphere with the same volume as the ROI (and therefore causes Roundness > 1).
Example in lung1 testcase:
* SA (pyradiomics) 782.24
* SA (SimpleITK) 576.06
* SA of a sphere with volume 1361.20: 593.97