STYL: Define default settings in separate function
Default setting for extraction are needed at initialization or when a parameter file is loaded. To prevent bugs due to inconsistencies in default settings defined at these locations, define the default settings in a separate function which is called when appropriate. Declare this function as a class method to enable easy usage outside the feature extractor module (e.g. when using feature classes directly).

N.B. Feature class specific settings and filter specific settings are NOT included here. These are defined in the feature classes and in `imageoperations`, respectively.