MATH: Change default value of voxelArrayShift
Set the default value to 0, which corresponds to most implementations of Energy, Total Energy and Root Mean Squared (RMS). This prevents users from accidentally using a too large array shift.

Update the documentation, to better define the usage of this parameter, with additional suggestions as to how to use it.

Update the formulas of features that incorporate this parameter to illustrate the effect this parameter has on the value.

Add documentation to explicitly mention the volume-confounded nature of these features, and indicate that the voxelArrayShift parameter influences the size of the effect of volume on the feature value.