STYL: Remove calls to 'print' and remove 'verbose' keyword
Restructure progress reporting in PyRadiomics to make use of logging package. Add a function `setVerbosity` to control verbosity level of the toolbox (level is specified as a logging level).
Replace calls to the `print` function by calls to `logger.info`. This also removes the need for the `verbose` keyword in the settings. When PyRadiomics is used in interactive python, a more verbose output can be obtained via the new `setVerbosity` function, defined at the module level. From the commandline, add an new optional argument (`--verbosity / -v`) to control verbosity (default level: WARNING, and when the argument is added without a value, level is set to INFO.