ENH: Add physical space check when resampling
Add a check that ensures that the ROI defined in the mask occupies a physical space that is contained within the physical space of the image. If this is not the case, log an error and cancel extraction for that combination.

Additionally, calculate the resampling grid using only the maskNode. This prevents unexpected behaviour of the defined resampling grid when image and mask with different geometry is used.

Standardize use of quote and double quote when defining strings in PyRadiomics (use single quote to identify a string, 3 double quotes for a docstring)