STYL: Add FAQ for image and mask of different geometry
Add FAQ explaining PyRadiomics behavior when using image and mask with different geometry.

Additionally, add an example script, usable from the commandline, that resamples a mask to a reference image. This allows users to force a mask to the same geometry as the image, so it can be used in PyRadiomics without resampling.