BUG: Error in geometry check during resampling
When checking if the ROI physical space is contained within the image physical space, direction is not taken into account when comparing in the physical space. Therefore, instead of bringing both the ROI and image bounds to the physical space, bring the ROI bounds first to the physical space and then to the image coordinate space. Then, compare those bounds with the minimum (always [-0.5, -0.5, -0.5]) and the maximum (image.GetSize() - 0.5) image bounds.

Additionally, add a tolerance to prevent machine precision errors (set at 1e-3).