ENH: Add pandas example
Add an example to show how pandas can be used to handle data returned by PyRadiomics.
Pandas is not needed to handle data inside PyRadiomics, and therefore does not have to be a dependency.
The result from pyradiomics can be converted to pandas by using it to instantiate a pandas Series.

The example shows how to use pandas for batchprocessing and storing the result, which users can adapt to suit their own needs.