ENH: Remove dependency for `tqdm`
Remove dependency for the tqdm package, as this is rarely used (only for GLCM and GLSZM in full-python mode).
Instead provide 3 variables which can hold function definitions used for instantiating and interacting with a progress bar.

In the helloRadiomics script and notebook, this changed functionality is illustrated with a commented out section in the script and a separate cell in the notebook. This section and cell assume the `tqdm` package is installed and use it to show a progressbar during calculation of the glcm matrix in full-python mode.