ENH: Add function to get test case
This function is available in the radiomics namespace and takes the test case name as an argument. Optionally the location of the repository can also be specified, in which case the test files of the repository are used.
If test files cannot be found, the are downloaded from github and stored in a temporary folder.

This function returns a tuple of 2 strings representing the paths to the (downloaded) image and segmentation files.