MATH: Remove SumVariance, rename SumVariance2
Remove the SumVariance feature and rename SumVariance2 to SumVariance. This is due to the fact that the original SumVariance in PyRadiomics was based on a literal interpretation of the definition in the Haralick paper, where it is possible that the reference to entropy instead of average was a typo (as in f7, it referred to f8 instead of f6).