ENH: Restructure Folders
Restructure folders to better catalog the scripts and jupyter notebooks not part of the pyradiomics core code.
Example scripts, notebooks and example parameter files are now stored in the directory tree under `examples`.
The `bin` folder now only contains some scripts useful for developers and for specific use cases. These include a script to generate and add a baseline for a new featureclass and a script to resample a mask to the image geometry.

Update the references to these files where necessary in the documentation.