DOCS: Document geometry errors and setting up logging
Add documentation on what goes wrong in case of geometry errors and how to fix it. Also include reference to this in the error message that is logged. Add the SimpleITK stacktrace as a debug-level log message in case of geometry errors.

Additionally, add more extensive documentation on how to set up logging (both to output and/or log file) in interactive and commandline usage of PyRadiomics.
Remove global PyRadiomics Function `debug`, the functionality exposed by this function is already covered by `setVerbosity`.