DOCS: Restructure documentation on customization
Add section on customizing the extraction in the "Usage" section. Here, specify what and how the extraction can be customized in PyRadiomics.
For this, move the documentation of the possible settings from the docstring in the featureextractor module to this new section. Also move the documentation that explains how the parameter file works.

Change the name of `self.kwargs` to `self.settings` in the feature extractor module, this better reflects what that variable holds.

 Make the distances setting only affect the GLCM (distance to neighbor should always be 1 in GLRLM and GLSZM).