ENH: Add case-specific labels in batchprocessing
Add functionality to allow users to specify different labels for different combinations in the input file for batchprocessing.
This is achieved through an additonal (optional) column ``Label``. Integer values in this column then take precedence over any other form
of label customization (e.g. specifying the label on the command line or in the parameter file).