BUG: Add directory check when downloading example data
When downloading example data, the feature visualization notebooks assumes a directory called `example_data` exists. However, if this is not the case, an error is thrown when trying to download the example_data zip file. Add a check to determine if the directory exists and if not, create a directory.

Fixes #276