Revert 280 / 280
Reverts #280 / #281, as this change introduces some confusion into the what goes into `**kwargs` for initialization of featureextractor (or one of the feature classes).

`**kwargs` only contains settings (i.e. one of the three categories of customization). PRs #280 and #281 included `inputImages` (one of the other categories) to enable customizing the enabled input image types in the slicer extension.
However, the functionality to customize this is already present in featureextractor through function calls (such as `enableAllInputImages()`). Therefore, revert these PRs and update the slicer extension to make use of this API.