STYL: Revise parameter naming
Make distinction between different categories of parameters more explicit, and rename categories to better reflect their definition.

PyRadiomics can be customized via 3 categories of *parameters*:
1. "imageType": specifies filters / image types that can be used to extract features from
2. "featureClass": specifies which feature classes and features should be extracted
3. "setting": specifies all setting to customize how the extraction is done.

In PyRadiomics, "parameter" refers to one or more customization categories and "setting" refers explicitly to type 3 customization.

Furthermore, document that only type 3 parameters can be provided at initialization when not using the parameter file (type 1 and 2 have to be changed using dedicated functions in featureextractor (type 1 and 2) or the individual feature classes (type 2).

Especially for type 1, this includes renaming of various variables, including the name used in the parameter file ("imageType" in stead of "inputImage").

Finally, expand documentation on making the parameter file, including an example directly in the documentation (instead of just referring to the repository).