ENH: Add resegmentation
Add functionality to resegment the mask prior to calculation of features (ony applies to firstorder / texture, no resegmentation possible prior to shape feature calculation). Introduces new parameter `resegmentRange`, a list of 2 floats which define the lower and upper threshold, respectively. Values outside the range specified are removed from the mask. When the resultant size is too small (parameter `minimumROISize`, default 1), an error is logged and the original mask is used. A value of None (the default value) for the parameter `resegmentRange` disables resegmentation.