Remove Sum Variance feature
Sum Variance is mathematically equal to Cluster Tendency, therefore remove Sum Variance from PyRadiomics.
Add a new section to the documentation that lists removed features with the reason why this was done (here including the proof that Sum Variance and Cluster Tendency are equal).

Additionally, Sum Average is very similar to Joint Average (if and only if the GLCM is symmetrical, Sum Average = 2 * Joint Average). As there is a difference possible between these features, but not in the case of 'default' settings, Sum Average is not removed, but disabled in the default parameter files.