ENH: Add additional checks to batch processing script
Add additional checks to ensure required columns are present and have a value for each case. If columns are missing, log an error and exit. If cells in required column have missing values (empty strings), log an error and skip to next case.

Also update the image and mask filepath if the path is relative, as it is assumed to be relative to the input file, not the current working directory.

Finally replace `os.path.exists` with `os.path.isfile` in `featureextractor.py`, as a directory is also an existing path, but should not be used during attempt to load using SimpleITK.