DOCS: Update documentation in base.py
Clarify that variables mentioned in the base class are variables instantiated by the base class, not possible settings for customization (also add redirect to customization documentation).

Update the documentation to reflect changes made to the structure of the base class, more specifically the initialization of a feature class.

Add documentation for the "Label" parameter to the customization documentation.