ENH: Add pre-cropping
When resampling is enabled, image and mask are pre-cropped onto the bounding box (with additional padding). This speeds up the resampling process and the application of filters. Moreover, it can significantly reduce the memory footprint of the extraction, especially when the images are large, but contain only small ROIs.
Add functionality to also pre-crop the image when no resampling is applied. This way, the advantages when applying filters on a pre-cropped image and mask are also available when extracting without resampling.

Similar as pre-cropping when resampling, additional padding is available (parameter `padDistance`), which is ensure to not exceed the original image bounds.