STYL: Re-add duplicate features as deprecated features
Re-add the feature calculation functions for the duplicate features, but instead of returning a result, raise a `DeprecationWarning` exception, which
is handled in the base class. This will not add the deprecated feature to the returned results, but log a warning to notify the user of the duplicity.

Additionally, for features with certain important assumptions, add warning messages to the log when these assumtions are not met:
- Joint Average, if the GLCM is not symmetrical
- Sum of Squares, if the GLCM is not symmetrical
- Sum Average, if the GLCM IS symmetrical (as it is then equal to 2 * Joint Average)