ENH: Add "deprecated" decorator
Add "deprecated" decorator to mark (feature) functions as being deprecated. This will add an `_is_deprecated` attribute to the function object (value will be `True`). When a feature function has this attribute (and set to `True`), this feature will not be enabled by default (i.e. when no individual features for that class are specified, or when all classes are enabled, again, without specifying individual features). These features can still be enabled by specifying them as individual features.

Some features are deprecated, but also return a deprecation warning. This is also included in the documentation.