BUG: Prevent testing of deprecated features
Unit testing determined test to run by checking all features that are contained in the feature class, but this also includes the deprecated features.

Change test case generation to use the list of features for which there is the baseline (in current class and current test case). Additionally, check if these baseline features include tests for all non-deprecated features.

Change the `getFeatureNames` function in `base.py` to generate a dictionary instead of a list: the key is the feature name, and the value is a boolean indicating if the feature is deprecated (`True`).