BUG: Fix error handling for image/mask mismatch
In `imageoperations.checkMask` and error is thrown if the image and mask do not align or contain tolerance errors. The error message is caught and parsed to provide users with a more insightful message providing potential solutions. However, this was done by evaluating the `message` attribute of the RunTimeError thrown by SimpleITK. This attribute is present in python 2, but not in python 3.
Instead, use `args[0]` attribute, which contains the message, and is present in both python 2 and 3.