ADD: Add PyRadiomics Model schema validation
Update the PyKwalify validation in PyRadiomics to also allow validation of PyRadiomics model files.

Add an example model file to explain the structure.

Update requirements.txt to enforce PyKwalify version != 1.6.0, as this version contains a bug that breaks the usage of partial schemas, which is utilized to allow validation of the two use cases (parameter file and model file)

Because partial schemas are used, load the schemas outside of pykwalify and pass them during validation as a dict.

Update the testParams script to allow testing of both use cases.