ADD: Add voxel-based extraction
Add a voxel based extraction, which outputs a dictionary of SimpleITK images instead of float values. Each voxel in the returned image represent the respective feature value for that region.

The local region can be defined as 'masked' or 'unmasked'. In unmasked mode, the region is defined as a cube with radius `kernelRadius` around the voxel: the kernel. If part of the kernel is outside of the image bounds, those parts are excluded.
In masked mode, the kernel is defined similar to unmasked mode, with additional exclusion of voxels that are outside the masked area (ROI).

This change also slightly changes the moment of several instantiation. In previous version, the the texture matrices in glcm, glrlm, etc. were calculated during initialization of the class. However, because for the voxel-based extraction they have to be calculated separately for each voxel, this functionality has been moved to the top of the `_calculateFeatures` (which runs when the features are acutally calculated).

N.B. This changes the API to the abstract feature class.
Features are now extracted by a call to `execute` instead of `calculateFeatures`. In addition to a name change, `execute` also directly returns the result (similar to usage of the feature extractor class). As before, calculated results remain available in the class under the `featureValues` attribute.

Adds a new section to the parameter file to allow customization of voxel-based extraction specific settings.

TODO:
 - Update examples to reflect API change
 - Add functionality in feature extractor module to make voxel based extraction available
 - Write new examples to reflect new functionality