TEST: Add wavelet test
Add a test to check if precropping and different size input images still produce the same output for the wavelet filter.
This test uses 2 new test cases that have been added to the 1.0.0 release of PyRadiomics and enabled for download by `getTestCase()`.
Additionally ensures temporal consistency by using a baseline file that defines the 1D-array of voxels in the ROI for each wavelet level (This array can be obtained by slicing the image array with a boolean cast of the mask array).

The baseline is generated using the large (64x64x64) image without pre cropping. The function to generate this baseline is added to `add_baseline.py`.

Finally, apply some style changes to the other test files to make them more consistent with pyradiomics style.

Supersedes #348.