ADD: Add customizable initialization value for voxel based extraction
Add optional customization of the initialization value for the output arrays when calculating features voxel based. By default this value is 0, but any float is allowed.

Additionally, change the name of the variable holding the customization settings in the feature classes from `kwargs` to `settings`. This better reflects what the variable holds and is similar to the variable name in the featureextractor module.