ENH: Update the getTestCase() function
Test data is now available als binary data attached to the v1.0 release. Update the `getTestCase()` function to retrieve images from this location (does not use the `data` directory anymore) and update the use of this function in the examples and testing scripts.
If data is available locally, this can still be used by passing it to the `getTestCase()` function, but it is not used in the repository anymore (allows for removal of the test cases from the repository itself to speed up cloning/downloading. See also #60.