BUG: Fix machine precision errors in eigen value calculation
If the segmentation in the dataset represents a 2D segmentation, the smallest of the 3 eigenvectors should be 0.
But because a transform to physical coordinates (taking into account rotation and scaling), this can actually compute very small (~1e-16) negative numbers.

This in turn causes an error in numpy.sqrt. This commit implements a 3-fold fix.

1) Because we do not need the eigen vectors, we can also calculate the eigenvalues using just the scaled coordinates, without rotation. In case of a 2D segmentation, this then results in correct calculation of 0 for the smallest eigen value.
2) After computation of eigen values. They are checked if they are negative, but > -1e-10. If that is the case, a warning is logged and they are rounded to 0.
3) Add additional checks to ensure the eigenvectors are not < 0 (effectively < -1e-10) in the feature functions. If that is the case, a warning is logged and nan is returned.