ADD: Add 3D implementation of Local Binary Pattern
Uses spherical harmonics implemented in `scipy` and defines an icosphere (implemented in `trimesh`) to sample neighbours. Derived images are returned for each level, with an additional derived image for the kurtosis.

To reduce computation time/cost, LBP is only calculated for voxels included in the segmentation. To implement this, the signature
of a filter function is changed (in addition to the image, the mask is also passed to the function). Update the examples, feature extractor and
documentation accordingly.