ENH: Further update getTestCase() function
No further changes to the interface are made.
However, if `dataDirectory` is specified and no test cases can be found there, it subsequently downloads testcases into that directory.
This allows users to get test cases easily, but store them locally (not in a temporary directory). If no directory is supplied, the temporary directory is used.
Attempts to create the directory if they don't already exist.