ENH: Switch to C generated angles
Add a function in C to generate angles and implement this directly in the C functions to calculate the texture matrices. As the angles are sometimes needed to determine weights to apply to the matrices, generated angles are also returned by the C function; i.e. a tuple as (matrix, angles).