ENH: Update/Simplify C functions for shape
Ensure consistency between C functions in shape and texture, move common functionality to separate function (various checks to ensure validity of input). Also remove use of Deprecated Numpy API functions and define the constant to enforce this.

Surface Area
Simplify function by removing interpolation (also removes the need for the edgeTable). This can be done as the point on the edge is always 0.5 (corner is always either 0 or 1 as it represents a binary mask). Apply a similar fix to the python calculation of surface area.

Diameters
Implement a simplified and customized function to generate angles, which are needed to find all border voxels.
Additionally, skip checking unnecessary indices (as mask is padded with 0's).