ENH: Update Matrices test script
Remove Python calculation from `test_cmatrices.py`. Instead, compare to baseline files (binary numpy array files).
Additionally, rename the script to `test_matrices.py`.

Update and move the script to add a new baseline to the tests folder, so that it can utilize the testUtils module
during generation of a new baseline. The updates now include the generation of a baseline matrix file if the class
contains the appropriate function (`_calculateCMatrix`).