ENH: Check if Matrix arrays are allocated correctly
If size is too big or the computer is unable to allocate memory for the output texture matrices, the function (PyArray_SimpleNew) returns NULL. Check for this and throw a python runtime error if that is the case.

Replace the large number with `INT_MAX` in the size check for increased readability.