ENH: Add publishing to PyPi to continuous integration
Adds a deploy step for all 3 ci-servers used by pyradiomics (supporting Linux, Windows and Mac OSX).
The deployment step is only run for tagged releases, i.e. when a build is triggered for a tag.
Furthermore, this tag is checked against a regex expression, to restrict deployment only to releases (e.g. v1.0, v2.1.3) and/or release candidates (e.g. v1.0-rc1, v2.1.3-rc12).

Authentication to the PyPi server is done using (private) environment variables set in the accounts on the various CI-servers. These variables will be only available to builds on branches/tags in the repository that is linked to the CI-account (i.e. not available for forked builds).

Finally, the current deployment step publishes to the test server, rather than the live server. This is done for testing of the workflow. Prior to integration into the master this will be rectified.