ENH: Simplify computation of various GLCM features
Inverse Difference, Inverse Difference Moment, Inverse Difference Normalized, Inverse Difference Moment Normalized and Inverse Variance all use a |i=j| element, which is equal to the k_diff. This enables the use of p_x-y (`pxSuby`) and k_diff (`kValuesDiff`), which are precomputed in `_calculateCoefficients`. Note that for Inverse Variance, the first element (k = 0) is skipped, in accordance with the constraint of i not equal to j in the formula.

As `pxSuby` contains fewer elements and dimensions than `P_glcm`, replacing `P_glcm` by `pxSuby` and `numpy.abs(i-j)` by `kValuesDiff` simplifies the function, and less computations are required.

Update the feature formulas in the documentation accordingly.

The final calculated result should be identical (within machine precision), therefore the baseline does not need to be updated.