ENH: Add script to create new tests
This script takes a configuration file, a test name (must be unique) and the test case to use. It then generates the output and stores this in the baseline files.

Update the test_features script to allow for tests that do not test all available features. (testing still fails if there are feature classes that are not tested).
When a feature class is tested, all features must be enabled.