BUG: Fix memory leak in C extensions
Since C functionality is implemented for the calculation of the angles, texture matrix calculation returns a tuple of the matrix and the angles.
However, this uses Py_BuildValue with specifier "O", which increases the reference count, and thereby introduces a memory leak.

Fix this by replacing "O" by "N", which has the same effect for the returned values, but does not increase the reference count.

Moreover, the angles array is only needed for GLCM and GLRLM matrices. Therefore, only return the angles array in these 2 functions and free the array in the other functions.